+++ src/degeneration/linear.cc
@@ -4,10 +4,10 @@
 namespace healthcare {
 namespace degeneration {
 
-Linear::Linear(double degen_intercept)
+Linear::Linear(float degen_intercept)
     : degen_intercept_(degen_intercept), degen_slope_(0) {}
 
-Linear::Linear(double degen_intercept, double degen_slope)
+Linear::Linear(float degen_intercept, float degen_slope)
     : degen_intercept_(degen_intercept), degen_slope_(degen_slope) {}
 
 int Linear::GetDegeneration(int age, int health) const {
